requirements.txt
websockets~=11.0.3
websocket-client~=1.5.3
pandas~=2.0.2
matplotlib~=3.7.1